* add in BaseLayout method for adding Virus * move all dto in DTO pack * add in VirusService method for getting List viruses * add showing in BaseController
