+++ src/main/resources/View/BaseLayout.fxml
@@ -1,11 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
 <?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
+<?import javafx.scene.layout.AnchorPane?>
 <AnchorPane prefHeight="434.0" prefWidth="767.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.math.epidemic.Controller.BaseController">
    <children>
       <TabPane layoutY="-1.0" prefHeight="434.0" prefWidth="767.0" tabClosingPolicy="UNAVAILABLE">
@@ -27,7 +23,8 @@
                               <TableColumn fx:id="cureVirusColumn" prefWidth="92.0" text="К.С. излечения" />
                           </columns>
                         </TableView>
-                        <Button layoutX="253.0" layoutY="366.0" mnemonicParsing="false"  text="Добавить" />
+                         <Button onAction="#onClickAdd" layoutX="253.0" layoutY="366.0" mnemonicParsing="false"
+                                 text="Добавить"/>
                         <Button layoutX="384.0" layoutY="366.0" mnemonicParsing="false"  text="Удалить" />
                      </children>
                   </AnchorPane>
